✨ Rename ifcondition
